0.6
Added hooks to UI in windows forms for enemy statistics. Added an enemy
class to handle the second party. Added enemy attacks. No AI yet. Added
ability to change current unit. Not available to the user yet.  Added
the ability for the player to lose. The enemy is undying for now