0.7
Added completed save and load functionality. Changed parties to be level
50 and have appropriate stats to match. The 'level' property in Unit.cs
has functionality to provide the level of a unit based on its exp, but
it is one level off for some reason. Added Combat Log functionality.
Added a combat log to the designer. Fixed a possible error with FSM
calling 'base()' instead of 'this()'. Added display for which unit is
currently fighting in a party by use of radio buttons in a group box.
Fixed error with tooltips overlapping and abilities hanging over from
previous units. Changed 's_Self' to be protected in the 'Singleton'
class. This may not be correct...

Added Unity project. Changed gitignore to match. changed settings to be
version control ready. Changed name of repository to reflect the fact
that this is the assessment.

Reorganized folder structure for unity project